+++ src/plugins/news/services/ArticleSubscriber.js
@@ -8,18 +8,14 @@ import getRequest from '@triniti/cms/plugins/pbjx/selectors/getRequest';
 import isCollaborating from '@triniti/cms/plugins/raven/selectors/isCollaborating';
 import resolveSchema from '@triniti/cms/utils/resolveSchema';
 import ArticleV1Mixin from '@triniti/schemas/triniti/news/mixin/article/ArticleV1Mixin';
-import swal from 'sweetalert2';
 import camelCase from 'lodash/camelCase';
 import get from 'lodash/get';
-import isEqual from 'lodash/isEqual';
 import isString from 'lodash/isString';
 import isUndefined from 'lodash/isUndefined';
-import { arrayPush, arrayRemoveAll, change, getFormAsyncErrors, getFormMeta, getFormSubmitErrors, getFormSyncErrors, getFormValues } from 'redux-form';
+import { arrayPush, arrayRemoveAll, change, getFormMeta, getFormValues } from 'redux-form';
 
 import formValues from '../utils/formValues';
 import { formNames, formRules } from '../constants';
-
-const createErrorMessage = (error, field) => (error.message ? `${field}: ${error.message}` : `${field}: an error occurred`);
 
 export default class ArticleSubscriber extends EventSubscriber {
   constructor() {
@@ -29,10 +25,14 @@ export default class ArticleSubscriber extends EventSubscriber {
     this.onArticleUpdated = this.onArticleUpdated.bind(this);
     this.onClearSubmitErrors = this.onClearSubmitErrors.bind(this);
     this.onInitForm = this.onInitForm.bind(this);
-    this.onSetSubmitFailed = this.onSetSubmitFailed.bind(this);
-    this.onSetSubmitSucceeded = this.onSetSubmitSucceeded.bind(this);
     this.onSubmitForm = this.onSubmitForm.bind(this);
     this.onValidateForm = this.onValidateForm.bind(this);
+  }
+
+  onClearSubmitErrors(event) {
+    const store = event.getRedux();
+    const { form } = store.getState();
+    formValues.set(form[formNames.ARTICLE]);
   }
 
   /**
@@ -181,10 +181,6 @@ export default class ArticleSubscriber extends EventSubscriber {
       return;
     }
 
-    // as much as possible, wrap each node update operation here with a try catch
-    // so we don't stop the form when 'FORCE' submitting. also at the same time, catch
-    // and invalidate the form when doing a regular submit
-
     [
       'allow_comments',
       'amp_enabled',
@@ -195,87 +191,53 @@ export default class ArticleSubscriber extends EventSubscriber {
       'smartnews_enabled',
       'title',
     ].forEach((fieldName) => {
-      try {
-        const value = data[camelCase(fieldName)];
-        node.set(fieldName, isUndefined(value) ? null : value);
-      } catch (error) {
-        console.error(`${fieldName}: `, error);
-        formEvent.addError(camelCase(fieldName), createErrorMessage(error, fieldName));
-      }
+      const value = data[camelCase(fieldName)];
+      node.set(fieldName, isUndefined(value) ? null : value);
     });
 
     [
       'apple_news_revision',
       'apple_news_share_url',
     ].forEach((fieldName) => {
-      try {
-        const value = get(data, camelCase(fieldName), null) || null;
-        if (node.has(fieldName) || value !== null) {
-          node.set(fieldName, value);
-        }
-      } catch (error) {
-        console.error(`${fieldName}: `, error);
-        formEvent.addError(camelCase(fieldName), createErrorMessage(error, fieldName));
+      const value = get(data, camelCase(fieldName), null) || null;
+      if (node.has(fieldName) || value !== null) {
+        node.set(fieldName, value);
       }
     });
 
-    try {
-      if (node.has('apple_news_id') || get(data, 'appleNewsId')) {
-        const value = get(data, 'appleNewsId');
+    if (node.has('apple_news_id') || get(data, 'appleNewsId')) {
+      const value = get(data, 'appleNewsId');
+      try {
         const identifier = isString(value) ? UuidIdentifier.fromString(value) : value;
         node.set('apple_news_id', identifier);
+      } catch (e) {
+        node.clear('apple_news_id');
       }
-    } catch (error) {
-      node.clear('apple_news_id');
-
-      console.error('apple_news_id: ', error);
-      formEvent.addError('appleNewsId', createErrorMessage(error, 'apple_news_id'));
     }
 
-    try {
-      node.set('image_ref', data.imageRef ? NodeRef.fromString(data.imageRef) : null);
-    } catch (error) {
-      console.error('image_ref: ', error);
-      formEvent.addError('imageRef', createErrorMessage(error, 'image_ref'));
-    }
+    node.set('image_ref', data.imageRef ? NodeRef.fromString(data.imageRef) : null);
 
-    try {
-      node.clear('slotting');
-      (data.slotting || []).forEach(({ key, value }) => {
-        if (key.label && value) {
-          node.addToMap('slotting', key.label, +value);
-        }
-      });
-    } catch (error) {
-      console.error('slotting: ', error);
-      formEvent.addError('slotting', createErrorMessage(error, 'slotting'));
-    }
-
-    try {
-      node.set('classification', get(data, `${camelCase('classification')}.value`) || null);
-    } catch (error) {
-      console.error('classification: ', error);
-      formEvent.addError('classification', createErrorMessage(error, 'classification'));
-    }
-
-    try {
-      node.clear('related_article_refs');
-      if (typeof data.relatedArticleRefs !== 'undefined') {
-        node.addToList('related_article_refs', data.relatedArticleRefs);
+    node.clear('slotting');
+    (data.slotting || []).forEach(({ key, value }) => {
+      if (key && key.label && value) {
+        node.addToMap('slotting', key.label, +value);
       }
-    } catch (error) {
-      console.error('related_article_refs: ', error);
-      formEvent.addError('relatedArticleRefs', createErrorMessage(error, 'related_article_refs'));
+    });
+
+    [
+      'classification',
+    ].forEach((fieldName) => {
+      node.set(fieldName, get(data, `${camelCase(fieldName)}.value`) || null);
+    });
+
+    node.clear('related_article_refs');
+    if (typeof data.relatedArticleRefs !== 'undefined') {
+      node.addToList('related_article_refs', data.relatedArticleRefs);
     }
 
-    try {
-      node.clear('author_ref');
-      if (data.authorRefs && data.authorRefs.length) {
-        node.set('author_ref', data.authorRefs[0]);
-      }
-    } catch (error) {
-      console.error('author_ref: ', error);
-      formEvent.addError('authorRefs', createErrorMessage(error, 'author_refs'));
+    node.clear('author_ref');
+    if (data.authorRefs && data.authorRefs.length) {
+      node.set('author_ref', data.authorRefs[0]);
     }
   }
 
@@ -413,68 +375,9 @@ export default class ArticleSubscriber extends EventSubscriber {
     }
   }
 
-  onClearSubmitErrors(event) {
-    const store = event.getRedux();
-    const state = store.getState();
-    const form = formNames.ARTICLE;
-    const errors = {
-      ...getFormSubmitErrors(form)(state),
-      ...getFormSyncErrors(form)(state),
-      ...getFormAsyncErrors(form)(state),
-    };
-    formValues.set(getFormValues(form)(state), errors);
-  }
-
-  onSetSubmitFailed(event) {
-    formValues.clear();
-  }
-
-  async onSetSubmitSucceeded(event) {
-    const store = event.getRedux();
-    const state = store.getState();
-    const locallyStoredValues = formValues.get();
-    if (!Object.keys(locallyStoredValues).length) {
-      return;
-    }
-
-    const submittedValues = getFormValues(formNames.ARTICLE)(state);
-    const serializedValues = JSON.parse(JSON.stringify(submittedValues));
-
-    // localstorage fields that have different values against the submitted values
-    // are invalid. this is because we revert any invalid values back to their initial values
-    // during force submit.
-    const invalidFields = Object.keys(locallyStoredValues)
-      .filter((fieldName) => !isEqual(locallyStoredValues[fieldName], serializedValues[fieldName]));
-
-    formValues.clear();
-
-    if (!invalidFields.length) {
-      return;
-    }
-
-    const result = await swal.fire({
-      title: `${invalidFields.length} unsave invalid fields. Want to keep the values and retry?`,
-      html: `<strong>${invalidFields.sort().join(', ')}</strong>`,
-      showCancelButton: true,
-      confirmButtonText: 'Yes!',
-      cancelButtonText: 'No',
-      confirmButtonClass: 'btn btn-primary',
-      cancelButtonClass: 'btn btn-secondary',
-      reverseButtons: true,
-    });
-
-    if (result.value) {
-      invalidFields.forEach((field) => {
-        store.dispatch(change(formNames.ARTICLE, field, locallyStoredValues[field]));
-      });
-    }
-  }
-
   getSubscribedEvents() {
     return {
       '@@redux-form/CLEAR_SUBMIT_ERRORS': this.onClearSubmitErrors,
-      '@@redux-form/SET_SUBMIT_FAILED': this.onSetSubmitFailed,
-      '@@redux-form/SET_SUBMIT_SUCCEEDED': this.onSetSubmitSucceeded,
       'triniti:news:mixin:article.init_form': this.onInitForm,
       'triniti:news:mixin:article.submit_form': this.onSubmitForm,
       'triniti:news:mixin:article.validate_form': this.onValidateForm,